+++ app/src/main/java/com/orogersilva/coloradissimo/di/scope/ActivityScope.java
@@ -1,0 +1,14 @@
+package com.orogersilva.coloradissimo.di.scope;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+import javax.inject.Scope;
+
+/**
+ * Created by orogersilva on 2/1/2016.
+ */
+@Scope
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ActivityScope {
+}
