+++ app/src/main/java/com/orogersilva/coloradissimo/interfaces/Validation.java
@@ -1,0 +1,13 @@
+package com.orogersilva.coloradissimo.interfaces;
+
+/**
+ * Created by orogersilva on 1/30/2016.
+ */
+public interface Validation {
+
+    // region SIGNATURES
+
+    void validate();
+
+    // endregion
+}
