+++ app/src/main/java/com/orogersilva/coloradissimo/model/ColoradissimoDatabase.java
@@ -1,0 +1,17 @@
+package com.orogersilva.coloradissimo.model;
+
+import com.raizlabs.android.dbflow.annotation.Database;
+
+/**
+ * Created by orogersilva on 1/30/2016.
+ */
+@Database(name = ColoradissimoDatabase.NAME, version = ColoradissimoDatabase.VERSION)
+public class ColoradissimoDatabase {
+
+    // region CONSTANTS
+
+    public static final String NAME = "ColoradoFanatico";
+    public static final int VERSION = 1;
+
+    // endregion
+}
