+++ app/src/main/java/com/orogersilva/coloradissimo/util/StringUtils.java
@@ -1,0 +1,16 @@
+package com.orogersilva.coloradissimo.util;
+
+/**
+ * Created by orogersilva on 1/30/2016.
+ */
+public class StringUtils {
+
+    // region STATIC METHODS
+
+    public static boolean isNullOrEmpty(String str) {
+
+        return str == null || str.isEmpty();
+    }
+
+    // endregion
+}
