--- app/src/main/java/com/orogersilva/coloradofanatico/di/component/AppComponent.java
@@ -1,22 +1,0 @@
-package com.orogersilva.coloradofanatico.di.component;
-
-import com.orogersilva.coloradofanatico.di.module.PersistenceModule;
-import com.orogersilva.coloradofanatico.model.FanModel;
-
-import javax.inject.Singleton;
-
-import dagger.Component;
-
-/**
- * Created by orogersilva on 1/30/2016.
- */
-@Singleton
-@Component(modules = {PersistenceModule.class})
-public interface AppComponent {
-
-    // region INJECTORS
-
-    void inject(FanModel fanModel);
-
-    // endregion
-}
