--- app/src/main/java/com/orogersilva/coloradofanatico/interfaces/Validation.java
@@ -1,13 +1,0 @@
-package com.orogersilva.coloradofanatico.interfaces;
-
-/**
- * Created by orogersilva on 1/30/2016.
- */
-public interface Validation {
-
-    // region SIGNATURES
-
-    void validate();
-
-    // endregion
-}
