--- app/src/main/java/com/orogersilva/coloradofanatico/model/ColoradoFanaticoDatabase.java
@@ -1,17 +1,0 @@
-package com.orogersilva.coloradofanatico.model;
-
-import com.raizlabs.android.dbflow.annotation.Database;
-
-/**
- * Created by orogersilva on 1/30/2016.
- */
-@Database(name = ColoradoFanaticoDatabase.NAME, version = ColoradoFanaticoDatabase.VERSION)
-public class ColoradoFanaticoDatabase {
-
-    // region CONSTANTS
-
-    public static final String NAME = "ColoradoFanatico";
-    public static final int VERSION = 1;
-
-    // endregion
-}
