--- app/src/main/java/com/orogersilva/coloradofanatico/util/StringUtils.java
@@ -1,16 +1,0 @@
-package com.orogersilva.coloradofanatico.util;
-
-/**
- * Created by orogersilva on 1/30/2016.
- */
-public class StringUtils {
-
-    // region STATIC METHODS
-
-    public static boolean isNullOrEmpty(String str) {
-
-        return str == null || str.isEmpty();
-    }
-
-    // endregion
-}
